Add React with Vite

+++ backend/controllers/product.controller.js
@@ -50,11 +50,15 @@ export const deleteProduct = async (req, res) => {
     // const id = req.params.id;
     const { id } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid product id" });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted" });
     } catch (error) {
         console.error("Error in Delete product: ", error.message);
-        res.status(404).json({ success: false, message: `Product not found: ${error.message}` });
+        res.status(500).json({ success: false, message: `Server error: ${error.message}` });
     }
 }
